GameEngineCore - AgentComponent events update
git-svn-id: http://hcm-lab.de/public/Horde3D/trunk@728 5ce291ac-9df0-446f-9e4f-d57731c4dda7

+++ Tools/GameEngine/src/GameEngineCore/GameEvent.h
@@ -1529,7 +1529,7 @@ public:
 		return m_returnI;
 	}
 
-	int getReturnFloat()
+	float getReturnFloat()
 	{
 		return m_returnF;
 	}
@@ -1547,7 +1547,7 @@ public:
 	const char* m_startNode;
 	const char* m_syncWord;
 	int m_returnI;
-	int m_returnF;
+	float m_returnF;
 };
 
 /**
